+++ NEWS.md
@@ -1,4 +1,4 @@
-# uom.dat 0.3.0.9000 (development version)
+# uom.dat 0.4.0 (2023-01-11)
 
 ## New features